move gist sync
